+++ src/main/jooqGenerated/org/xbery/artbeams/jooq/schema/indexes/Indexes.kt
@@ -10,8 +10,10 @@ import org.jooq.impl.Internal
 import org.xbery.artbeams.jooq.schema.tables.AuthCode
 import org.xbery.artbeams.jooq.schema.tables.Comments
 import org.xbery.artbeams.jooq.schema.tables.Media
+import org.xbery.artbeams.jooq.schema.tables.Orders
 import org.xbery.artbeams.jooq.schema.tables.Queue
 import org.xbery.artbeams.jooq.schema.tables.UserAccess
+import org.xbery.artbeams.jooq.schema.tables.Users
 
 
 
@@ -22,6 +24,9 @@ import org.xbery.artbeams.jooq.schema.tables.UserAccess
 val IDX_AUTH_CODE_USER_ID: Index = Internal.createIndex(DSL.name("idx_auth_code_user_id"), AuthCode.AUTH_CODE, arrayOf(AuthCode.AUTH_CODE.USER_ID), false)
 val IDX_COMMENTS_ENTITY_ID: Index = Internal.createIndex(DSL.name("idx_comments_entity_id"), Comments.COMMENTS, arrayOf(Comments.COMMENTS.ENTITY_ID), false)
 val IDX_MEDIA_FILENAME: Index = Internal.createIndex(DSL.name("idx_media_filename"), Media.MEDIA, arrayOf(Media.MEDIA.FILENAME), false)
+val IDX_ORDERS_ORDER_NUMBER: Index = Internal.createIndex(DSL.name("idx_orders_order_number"), Orders.ORDERS, arrayOf(Orders.ORDERS.ORDER_NUMBER), true)
 val IDX_QUEUE_EXPIRATION: Index = Internal.createIndex(DSL.name("idx_queue_expiration"), Queue.QUEUE, arrayOf(Queue.QUEUE.EXPIRATION_TIME), false)
 val IDX_QUEUE_NEXT_ACTION_TIME: Index = Internal.createIndex(DSL.name("idx_queue_next_action_time"), Queue.QUEUE, arrayOf(Queue.QUEUE.NEXT_ACTION_TIME), false)
 val IDX_USER_ACCESS_UNIQUE: Index = Internal.createIndex(DSL.name("idx_user_access_unique"), UserAccess.USER_ACCESS, arrayOf(UserAccess.USER_ACCESS.ACCESS_DATE, UserAccess.USER_ACCESS.IP, UserAccess.USER_ACCESS.USER_AGENT, UserAccess.USER_ACCESS.ENTITY_TYPE, UserAccess.USER_ACCESS.ENTITY_ID), true)
+val IDX_USERS_EMAIL: Index = Internal.createIndex(DSL.name("idx_users_email"), Users.USERS, arrayOf(Users.USERS.EMAIL), true)
+val IDX_USERS_LOGIN: Index = Internal.createIndex(DSL.name("idx_users_login"), Users.USERS, arrayOf(Users.USERS.LOGIN), true)