+++ src/main/jooqGenerated/org/xbery/artbeams/jooq/schema/tables/Orders.kt
@@ -6,8 +6,11 @@ package org.xbery.artbeams.jooq.schema.tables
 
 import java.time.Instant
 
+import kotlin.collections.List
+
 import org.jooq.Field
 import org.jooq.ForeignKey
+import org.jooq.Index
 import org.jooq.Name
 import org.jooq.Record
 import org.jooq.Schema
@@ -21,6 +24,7 @@ import org.jooq.impl.SQLDataType
 import org.jooq.impl.TableImpl
 import org.xbery.artbeams.common.persistence.jooq.converter.InstantConverter
 import org.xbery.artbeams.jooq.schema.DefaultSchema
+import org.xbery.artbeams.jooq.schema.indexes.IDX_ORDERS_ORDER_NUMBER
 import org.xbery.artbeams.jooq.schema.keys.CONSTRAINT_C3
 import org.xbery.artbeams.jooq.schema.tables.records.OrdersRecord
 
@@ -113,6 +117,7 @@ open class Orders(
 
     constructor(child: Table<out Record>, key: ForeignKey<out Record, OrdersRecord>): this(Internal.createPathAlias(child, key), child, key, ORDERS, null)
     override fun getSchema(): Schema? = if (aliased()) null else DefaultSchema.DEFAULT_SCHEMA
+    override fun getIndexes(): List<Index> = listOf(IDX_ORDERS_ORDER_NUMBER)
     override fun getPrimaryKey(): UniqueKey<OrdersRecord> = CONSTRAINT_C3
     override fun `as`(alias: String): Orders = Orders(DSL.name(alias), this)
     override fun `as`(alias: Name): Orders = Orders(alias, this)