+++ src/main/kotlin/org/xbery/artbeams/comments/domain/CommentState.kt
@@ -1,0 +1,11 @@
+package org.xbery.artbeams.comments.domain
+
+/**
+ * @author Radek Beran
+ */
+enum class CommentState {
+    /** Waiting for approval. */
+    WAITING_FOR_APPROVAL,
+    APPROVED,
+    REJECTED
+}