+++ src/main/kotlin/org/xbery/artbeams/comments/repository/CommentRepository.kt
@@ -5,6 +5,7 @@ import org.jooq.Field
 import org.jooq.Table
 import org.springframework.stereotype.Repository
 import org.xbery.artbeams.comments.domain.Comment
+import org.xbery.artbeams.comments.domain.CommentState
 import org.xbery.artbeams.comments.repository.mapper.CommentMapper
 import org.xbery.artbeams.comments.repository.mapper.CommentUnmapper
 import org.xbery.artbeams.common.assets.repository.AssetRepository
@@ -31,9 +32,21 @@ class CommentRepository(
             .orderBy(COMMENTS.CREATED.desc())
             .fetch(mapper)
 
-    fun findByEntityId(entityId: String): List<Comment> =
+    /**
+     * Finds approved comments for an entity (e.g. article id).
+     * @param entityId
+     * @return
+     */
+    fun findApprovedByEntityId(entityId: String): List<Comment> =
         dsl.selectFrom(table)
-            .where(COMMENTS.ENTITY_ID.eq(entityId))
+            .where(COMMENTS.ENTITY_ID.eq(entityId), COMMENTS.STATE.eq(CommentState.APPROVED.name))
             .orderBy(COMMENTS.CREATED)
             .fetch(mapper)
+
+    fun updateState(id: String, state: CommentState): Boolean {
+        return dsl.update(COMMENTS)
+            .set(COMMENTS.STATE, state.name)
+            .where(COMMENTS.ID.eq(id))
+            .execute() > 0
+    }
 }