+++ src/main/kotlin/org/xbery/artbeams/comments/repository/mapper/CommentMapper.kt
@@ -3,6 +3,7 @@ package org.xbery.artbeams.comments.repository.mapper
 import org.jooq.RecordMapper
 import org.springframework.stereotype.Component
 import org.xbery.artbeams.comments.domain.Comment
+import org.xbery.artbeams.comments.domain.CommentState
 import org.xbery.artbeams.common.access.domain.EntityKey
 import org.xbery.artbeams.common.assets.domain.AssetAttributes
 import org.xbery.artbeams.jooq.schema.tables.records.CommentsRecord
@@ -23,6 +24,7 @@ class CommentMapper : RecordMapper<CommentsRecord, Comment> {
                 modifiedBy = requireNotNull(record.modifiedBy)
             ),
             parentId = record.parentId,
+            state = CommentState.valueOf(requireNotNull(record.state)),
             comment = requireNotNull(record.comment),
             userName = requireNotNull(record.username),
             email = requireNotNull(record.email),