+++ src/main/kotlin/org/xbery/artbeams/comments/service/CommentService.kt
@@ -1,11 +1,12 @@
 package org.xbery.artbeams.comments.service
 
 import org.xbery.artbeams.comments.domain.Comment
+import org.xbery.artbeams.comments.domain.CommentState
 import org.xbery.artbeams.comments.domain.EditedComment
 import org.xbery.artbeams.common.context.OperationCtx
 
 interface CommentService {
-    fun findByEntityId(entityId: String): List<Comment>
+    fun findApprovedByEntityId(entityId: String): List<Comment>
 
     fun saveComment(
         edited: EditedComment,
@@ -16,4 +17,5 @@ interface CommentService {
 
     fun findComments(): List<Comment>
     fun deleteComment(id: String): Boolean
+    fun updateCommentState(id: String, state: CommentState): Boolean
 }